+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -1,5 +1,6 @@
 package ru.yandex.practicum.filmorate.service;
 
+import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.util.List;
@@ -21,4 +22,5 @@ public interface UserService {
 
     List<User> getMutualFriends(int userId, int otherId);
 
+    List<Film> getRecommendedFilms(int id);
 }
