+++ src/main/java/ru/yandex/practicum/filmorate/service/impl/FilmServiceImpl.java
@@ -10,7 +10,6 @@ import ru.yandex.practicum.filmorate.model.*;
 import ru.yandex.practicum.filmorate.service.FilmService;
 
 import java.util.*;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -148,42 +147,6 @@ public class FilmServiceImpl implements FilmService {
             throw new NotFoundException("Режиссер с id = " + directorId + " не найден");
         }
         return films;
-    }
-
-    @Override
-    public List<Film> getRecommendedFilms(int userId) {
-        if (userId < 0) {
-            throw new ValidationException("Идентификатор пользователя должен быть положительным числом");
-        }
-
-        // Фильмы, которые поставили лайк пользователь X, делавший запрос
-        List<Film> filmsLikedByUser = filmRepository.getFilmsLikedByUser(userId);
-
-        Set<Integer> filmsIdsLikedByUser = filmsLikedByUser.stream()
-                    .map(Film::getId)
-                    .collect(Collectors.toSet());
-
-        // Пользователи, которые поставили лайк те же самые фильмы, что и пользователь X
-        List<User> usersThatLikedSameFilms = userRepository.getUsersWithSameLikes(filmsIdsLikedByUser);
-        List<Integer> usersIds = usersThatLikedSameFilms.stream()
-                .map(User::getId)
-                .toList();
-        HashMap<Integer, List<Film>> foundUsersAllLikedFilms = filmRepository.getLikedFilmsByUsersIds(usersIds);
-
-
-        // Фильмы, которые не поставили лайк пользователь X, делавший запрос
-        List<Film> recommendedFilms = new ArrayList<>();
-
-        for (Map.Entry<Integer, List<Film>> entry : foundUsersAllLikedFilms.entrySet()) {
-            List<Film> likedByOtherUser = entry.getValue();
-
-            for (Film film : likedByOtherUser) {
-                if (!filmsLikedByUser.contains(film)) {
-                    recommendedFilms.add(film);
-                }
-            }
-        }
-        return recommendedFilms;
     }
 
 }
